requirements.txt
amqp==5.0.5
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
future==0.18.2
importlib-metadata==3.4.0
iso8601==0.1.13
kombu==5.0.2
pkg-resources==0.0.0
prompt-toolkit==3.0.14
pyserial==3.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
RPi.GPIO==0.7.0
serial==0.0.97
six==1.15.0
smbus==1.1.post2
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
